feat(utils) added ability to unconditionally cleanup
if you pass an argument to contrib/clean_running_apps.sh
then it will skip the prompt and do the cleanup unconditionally

Closes: #218
Approved by: jappavoo

+++ contrib/clean_running_apps.sh
@@ -1,11 +1,12 @@
 #!/bin/bash
 
 # Manual clean-up of docker networks and containers created by the EbbRT NodeAllocator
-
-read -p "Remove all of your ($USER) ebbrt networks and nodes. Continue [Y]es, [n]o? " c
-case "$c" in 
-  n|N ) exit 0;
-esac
+if [[ -z $1 ]]; then
+  read -p "Remove all of your ($USER) ebbrt networks and nodes. Continue [Y]es, [n]o? " c
+  case "$c" in 
+    n|N ) exit 0;
+  esac
+fi
 
 echo "Removing docker containers..."
 docker ps | grep ebbrt-$(id -u) | cut -d ' ' -f 1 | while read id; do docker kill $id; done
